Organized files and added setup.py to make the package ready for PYPI
