+++ historicalFinancialData/__init__.py
@@ -1,0 +1,12 @@
+"""
+historicalFinancialData
+
+A library for public companies' historical financial data reported on a quarterly basis.
+"""
+
+from . import main
+from historicalFinancialData.main import FinData
+
+__version__ = "0.1.0"
+__author__ = 'Daniel Mistrik'
+__credits__ = 'Data courtesy of the SEC (This project is not affiliated with, approved or vetted by the SEC)'
