+++ historicalFinancialData/exceptions.py
